backend/requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
requests==2.31.0
beautifulsoup4==4.12.2
pymongo==4.5.0
python-dotenv==1.0.0
apscheduler==3.10.4
lxml==4.9.3
httpx==0.25.0
email-validator==2.2.0
passlib[bcrypt]==1.7.4
